Change more global variables to atomics to ensure cross-thread synchronization. Move the direct register operations out of version.c and into saturnregisters.c, for better encapsulation. Use facade functions to encapsulate the complex logic.
Simplify FIFO depth initializers, eliminating a branch in the FIFO read function.

+++ sw_projects/P2_app/InHighPriority.c
@@ -45,7 +45,7 @@ void *IncomingHighPriority(void *arg)                   // listener thread
   uint16_t Word;
   int i;                                                // counter
   ESoftwareID FPGASWID;                                 // preprod/release etc
-  unsigned int FPGAVersion;                             // firmware version
+  uint16_t FPGAVersion;                             // firmware version
   bool PAEnable;
